backend/requirements.txt
# 4XL Backend Dependencies
# Real-ESRGAN Image Upscaler API

# Core framework
gradio>=4.44.0
spaces>=0.19.0

# PyTorch (CUDA) - HF Spaces will use appropriate version
torch>=2.0.0
torchvision>=0.15.0,<0.17.0

# Real-ESRGAN and dependencies
realesrgan>=0.3.0
basicsr>=1.4.2

# Image processing
numpy<2
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# Utilities
requests>=2.31.0
huggingface_hub>=0.23.0
